% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_topic_score.R
\name{calc_topic_score_sql}
\alias{calc_topic_score_sql}
\title{Calcuate the topic score from beta data table.}
\usage{
calc_topic_score_sql(df, word, beta)
}
\arguments{
\item{df}{data.frame. beta data table}

\item{word}{Character. word column.}

\item{beta}{numeric. beta scores for the words.}
}
\description{
Calcuate the topic score from beta data table.
}
\examples{
beta_df <- data.frame(dict = c("文本", "分析", "学习", "笔记"),
                      loading = c(0.1, 0.2, 0.3, 0.4))
calc_topic_score_sql(beta_df, dict, loading)
}
