% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_topic_score.R
\name{str_count_sql}
\alias{str_count_sql}
\title{Count the regular expression frequency.}
\usage{
str_count_sql(input = "x1", regexp = "测试", is_rename = TRUE,
  output = input)
}
\description{
Count the regular expression frequency.
}
\examples{
\dontrun{str_count_sql(input = "x1", regexp = "娴嬭瘯")}
}
